chore: refactor (wip)
