+++ backend/lambdas/offers/functions/initDb/init.sql
