+++ backend/lambdas/offers/package.json
@@ -13,6 +13,7 @@
     "typescript": "^5.6.3"
   },
   "dependencies": {
+    "aws-lambda": "^1.0.7",
     "aws-sdk": "^2.1692.0",
     "jwt-decode": "^4.0.0",
     "pg": "^8.13.1"
