+++ backend/lambdas/offers/tsconfig.json
@@ -5,7 +5,13 @@
       "target": "es6",
       "moduleResolution": "node",
       "sourceMap": true,
-      "outDir": "dist"
+      "outDir": "dist",
+      "paths": {
+        "@shared/*": [
+          "../../shared/*",
+          "./shared/*"
+        ]
+      }
     },
     "lib": ["es2015"],
     "include": ["./**/*.ts"],
